feat(mcp): add tools query_logs and get_rootcause_details
- added resource logs://docs/0 for summary of current opened
dlt document.
- modified resource fishbones://... to return a json summary.
